Updated for the RDB tables and log4j
