+++ nnd-data-poll-service/build.gradle
@@ -32,7 +32,7 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
     implementation 'com.microsoft.sqlserver:mssql-jdbc:12.2.0.jre11'
-
+    implementation 'net.logstash.logback:logstash-logback-encoder:7.4'
     testImplementation(platform('org.junit:junit-bom'))
     testImplementation('org.junit.jupiter:junit-jupiter')
     testImplementation 'org.junit.jupiter:junit-jupiter-engine'
