+++ nnd-data-poll-service/src/main/java/gov/cdc/nnddatapollservice/rdb/service/interfaces/IRdbDataHandlingService.java
@@ -1,0 +1,7 @@
+package gov.cdc.nnddatapollservice.rdb.service.interfaces;
+
+import gov.cdc.nnddatapollservice.exception.DataPollException;
+
+public interface IRdbDataHandlingService {
+    void handlingExchangedData() throws DataPollException;
+}
