--- nnd-data-poll-service/src/main/java/gov/cdc/nnddatapollservice/service/interfaces/IRdbDataHandlingService.java
@@ -1,12 +1,0 @@
-package gov.cdc.nnddatapollservice.service.interfaces;
-
-import gov.cdc.nnddatapollservice.exception.DataPollException;
-import gov.cdc.nnddatapollservice.rdb.dto.PollDataSyncConfig;
-
-import java.util.List;
-
-public interface IRdbDataHandlingService {
-    void handlingExchangedData(String tableName) throws DataPollException;
-    void persistRdbData(String tableName, String jsonData) throws DataPollException;
-    List<PollDataSyncConfig> getTableListFromConfig() throws DataPollException;
-}
